Create insertion_a_node
A Singly Linked List (LL) is a linear data structure where each element, known as a node, is connected to its next element using a pointer. Unlike arrays, the elements in a singly linked list are not stored in contiguous memory locations. Each node contains a data element and a pointer to the next node in the sequence. The list starts with a head node, which points to the first node, and ends with a tail node, which points to the last node. The tail node's pointer is null, indicating the end of the list.